avoid hook duplication and minor fixes
